requirements.txt
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
Cython==0.29.24
docutils==0.17.1
firebase-admin==5.0.3
google-api-core==2.1.0
google-api-python-client==2.24.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
httplib2==0.20.1
idna==3.2
Kivy==2.0.0
Kivy-Garden==0.1.4
kivymd @ https://github.com/kivymd/KivyMD/archive/master.zip
msgpack==1.0.2
multitasking==0.0.9
packaging==21.0
Pillow==8.3.2
proto-plus==1.19.3
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==2.4.7
requests==2.26.0
rsa==4.7.2
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.7
watchdog==2.1.6
